pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Kobina Brandon <kobinabrandon@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
loguru = "^0.7.2"
quixstreams = "^3.3.0"
websockets = "^14.1"
pydantic-settings = "^2.6.1"
geopy = "^2.4.1"
python-dotenv = "^1.0.1"
tqdm = "^4.67.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
